// types/game.ts
export type Game = {
  id: number;
  title: string;
  description: string;
  image: string;
  alt: string;
  buttonText: string;
  playLink: string;
};

export interface GameInterface {
  id: number;
  title: string;
  description: string;
  image: string;
  alt: string;
  buttonText: string;
  playLink: string;
};